ScriptableObject + cross-scene reference support
Added ScriptableObject version of the ToDo list. Works identical to the MonoBehaviour version, just isn't attached to a Scene.
Added ToDo lists to Create menus and "Cyan" heading to AddComponent menu
Added SceneReferencesHandler which is automatically added to a scene when required.
It's now possible to link objects to a task which are cross-scene, or scene-based on a prefab or ScriptableObject ToDo list. This has only been tested in 2020.1 currently.

How it works :
When dragging an object onto a task from a different scene, or from a scene when the ToDo list is a prefab/ScriptableObject, a SceneReferencesHandler object is created and added to the scene - hidden from the Hierarchy but still serialised in scene. The dragged object is registered with the handler, which assigns it a unique ID, adds it to a SerializableDictionary and saves the scene. The task can then reference and serialise the SceneAsset (in element.objectReference) and the ID (in element.objectReferenceID). When the task is visible in-editor, it will attempt to obtain the SceneReferenceHandler from the SceneAsset (if loaded, sidenote : won't work at runtime sadly, would need to save the scene name/path/buildIndex instead). With this it can obtain the object from the ID and show it in the object field. Fields which are serialised in this way are marked with a asterisk.